+++ system/info.go
@@ -3,7 +3,7 @@ package system
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
-//                         Copyright (c) 2022 ESSENTIAL KAOS                          //
+//                         Copyright (c) 2023 ESSENTIAL KAOS                          //
 //      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -12,22 +12,6 @@ import (
 	"bufio"
 	"os"
 	"strconv"
-)
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-// OS names
-const (
-	LINUX_ARCH      = "Arch"
-	LINUX_CENTOS    = "CentOS"
-	LINUX_DEBIAN    = "Debian"
-	LINUX_FEDORA    = "Fedora"
-	LINUX_GENTOO    = "Gentoo"
-	LINUX_RHEL      = "RHEL"
-	LINUX_SUSE      = "SuSe"
-	LINUX_OPEN_SUSE = "openSUSE"
-	LINUX_UBUNTU    = "Ubuntu"
-	DARWIN_OSX      = "OSX"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -138,13 +122,12 @@ type InterfaceStats struct {
 
 // SystemInfo contains info about a system (hostname, OS, arch...)
 type SystemInfo struct {
-	Hostname     string `json:"hostname"`     // Hostname
-	OS           string `json:"os"`           // OS name
-	Distribution string `json:"distribution"` // OS distribution
-	Version      string `json:"version"`      // OS version
-	Kernel       string `json:"kernel"`       // Kernel version
-	Arch         string `json:"arch"`         // System architecture (i386/i686/x86_64/etc...)
-	ArchBits     int    `json:"arch_bits"`    // Architecture bits (32/64)
+	Hostname string `json:"hostname"`  // Hostname
+	OS       string `json:"os"`        // OS name
+	Kernel   string `json:"kernel"`    // Kernel version
+	Arch     string `json:"arch"`      // System architecture (i386/i686/x86_64/etc…)
+	ArchName string `json:"arch_name"` // System architecture (386/686/amd64/etc…)
+	ArchBits int    `json:"arch_bits"` // Architecture bits (32/64)
 }
 
 // OSInfo contains info about OS
@@ -152,6 +135,7 @@ type OSInfo struct {
 	Name                  string `json:"name"`
 	PrettyName            string `json:"pretty_name"`
 	Version               string `json:"version"`
+	Build                 string `json:"build"`
 	VersionID             string `json:"version_id"`
 	VersionCodename       string `json:"version_codename"`
 	ID                    string `json:"id"`