+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 12.58.0
+
+* `[system]` Added system arch name to `SystemInfo`
+* `[system]` `Version` and `Distribution` info removed from `SystemInfo` (_use `OSInfo` instead_)
+* `[system]` `GetOSInfo` now works on macOS
+
 ### 12.57.1
 
 * `[progress]` Fixed bug with updating progress settings